Finished rough OOP version and added documentation / screenshot

+++ object_oriented_version/client/client.h
@@ -19,13 +19,14 @@
 
 class Client {
     private:
-        int port;
+        long port;
         char *serverIp;
     public:
-        Client(int port, char *serverIp) {
+        Client(long port, char* serverIp) {
             this->port = port;
             this->serverIp = serverIp;
         }
 
+        // Runs chat from the client end
         void clientConnect(void);
 };
